update module names
