+++ integration-tests/oauth2-reactive-spring/src/test/resources/versions.properties
