+++ integration-tests/oauth2-servlet-spring/src/test/resources/testRunner.yml
@@ -1,0 +1,84 @@
+#
+# Copyright 2021-Present Okta, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+implementation: com.okta.spring.tests.oauth2.SpringApplicationUnderTest
+scenarios:
+
+  code-flow-remote-validation:
+    command: com.okta.spring.tests.common.servlet.code.BasicRedirectCodeFlowApplication
+    args:
+      - --server.port=${applicationPort}
+      - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+      - --spring.security.oauth2.client.provider.okta.authorization-uri=https://localhost:${mockHttpsPort}/oauth2/default/v1/authorize
+      - --spring.security.oauth2.client.provider.okta.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+      - --spring.security.oauth2.client.registration.okta.client-id=OOICU812
+      - --spring.security.oauth2.client.registration.okta.client-secret=VERY_SECRET
+      - --spring.security.oauth2.client.registration.okta.scope=offline_access
+      - --spring.security.oauth2.client.registration.okta.redirect-uri={baseUrl}/authorization-code/callback
+      - --server.servlet.session.tracking-modes=cookie
+    protectedPath: /oauth2/authorization/okta
+
+  implicit-flow-local-validation:
+    command: com.okta.spring.tests.common.servlet.jwt.BasicJwtResourceServerApplication
+    args:
+    - --server.port=${applicationPort}
+    - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+    - --spring.security.oauth2.resourceserver.jwt.jwk-set-uri=https://localhost:${mockHttpsPort}/oauth2/default/v1/keys
+    # A client ID is NOT required, #218
+    - --server.session.trackingModes=cookie
+
+  implicit-flow-remote-validation:
+    command: com.okta.spring.tests.common.servlet.opaquetoken.BasicOpaqueTokenResourceServerApplication
+    args:
+      - --server.port=${applicationPort}
+      - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+      - --spring.security.oauth2.resourceserver.jwt.jwk-set-uri=https://localhost:${mockHttpsPort}/oauth2/default/v1/keys
+      - --spring.security.oauth2.resourceserver.opaque-token.client-id=OOICU812
+      - --spring.security.oauth2.resourceserver.opaque-token.client-secret=VERY_SECRET
+      - --spring.security.oauth2.resourceserver.opaque-token.introspection-uri=https://localhost:${mockHttpsPort}/oauth2/default/v1/introspect
+      - --server.session.trackingModes=cookie
+
+  oidc-code-flow-local-validation:
+    disabledTests:
+      # Spring does remote validation of the token
+      - invalidIssuerIdTokenJwtTest
+    command: com.okta.spring.tests.common.servlet.code.BasicRedirectCodeFlowApplication
+    args:
+      - --server.port=${applicationPort}
+      - --server.servlet.session.tracking-modes=cookie
+      - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+      - --spring.security.oauth2.client.provider.okta.authorization-uri=https://localhost:${mockHttpsPort}/oauth2/default/v1/authorize
+      - --spring.security.oauth2.client.provider.okta.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+      - --spring.security.oauth2.client.registration.okta.client-id=OOICU812
+      - --spring.security.oauth2.client.registration.okta.client-secret=VERY_SECRET
+      - --spring.security.oauth2.client.registration.okta.redirect-uri={baseUrl}/authorization-code/callback
+      - --spring.security.oauth2.client.registration.okta.scope=profile,email,openid
+      - --okta.oauth2.postLogoutRedirectUri=/you-are-logged-out
+    protectedPath: /oauth2/authorization/okta
+
+  pkce-code-flow-remote-validation:
+    command: com.okta.spring.tests.common.servlet.code.BasicRedirectCodeFlowApplication
+    args:
+      - --server.port=${applicationPort}
+      - --server.servlet.session.tracking-modes=cookie
+      - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+      - --spring.security.oauth2.client.provider.okta.authorization-uri=https://localhost:${mockHttpsPort}/oauth2/default/v1/authorize
+      - --spring.security.oauth2.client.provider.okta.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
+      - --spring.security.oauth2.client.registration.okta.client-id=OOICU812
+      - --spring.security.oauth2.client.registration.okta.scope=offline_access
+      - --spring.security.oauth2.client.registration.okta.redirect-uri={baseUrl}/authorization-code/callback
+      - --spring.security.oauth2.client.registration.okta.client-authentication-method=none
+    protectedPath: /oauth2/authorization/okta
