+++ integration-tests/oauth2-servlet-spring/src/test/resources/versions.properties
