+++ integration-tests/oauth2-servlet/src/test/resources/testRunner.yml
@@ -1,0 +1,73 @@
+#
+# Copyright 2017 Okta, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+implementation: com.okta.spring.tests.oauth2.SpringApplicationUnderTest
+scenarios:
+
+  code-flow-remote-validation:
+    command: com.okta.spring.tests.common.servlet.code.BasicRedirectCodeFlowApplication
+    args:
+    - --server.port=${applicationPort}
+    - --okta.oauth2.issuer=https://localhost:${mockHttpsPort}/oauth2/default
+    - --okta.oauth2.clientId=OOICU812
+    - --okta.oauth2.clientSecret=VERY_SECRET
+    - --okta.oauth2.scopes=offline_access
+    - --okta.oauth2.redirectUri=/authorization-code/callback
+    - --server.servlet.session.tracking-modes=cookie
+    protectedPath: /oauth2/authorization/okta
+
+  implicit-flow-local-validation:
+    command: com.okta.spring.tests.common.servlet.jwt.BasicJwtResourceServerApplication
+    args:
+    - --server.port=${applicationPort}
+    - --okta.oauth2.issuer=https://localhost:${mockHttpsPort}/oauth2/default
+    # A client ID is NOT required, #218
+    - --server.session.trackingModes=cookie
+
+  implicit-flow-remote-validation:
+    command: com.okta.spring.tests.common.servlet.opaquetoken.BasicOpaqueTokenResourceServerApplication
+    args:
+      - --server.port=${applicationPort}
+      - --okta.oauth2.issuer=https://localhost:${mockHttpsPort}/oauth2/default
+      - --okta.oauth2.clientId=OOICU812
+      - --okta.oauth2.clientSecret=VERY_SECRET
+      - --server.session.trackingModes=cookie
+
+  oidc-code-flow-local-validation:
+    disabledTests:
+    # Spring does remote validation of the token
+    - invalidIssuerIdTokenJwtTest
+    command: com.okta.spring.tests.common.servlet.code.BasicRedirectCodeFlowApplication
+    args:
+    - --server.port=${applicationPort}
+    - --okta.oauth2.issuer=https://localhost:${mockHttpsPort}/oauth2/default
+    - --okta.oauth2.clientId=OOICU812
+    - --okta.oauth2.clientSecret=VERY_SECRET
+    - --okta.oauth2.redirectUri=/authorization-code/callback
+    - --okta.oauth2.postLogoutRedirectUri=/you-are-logged-out
+    - --server.servlet.session.tracking-modes=cookie
+    protectedPath: /oauth2/authorization/okta
+
+  pkce-code-flow-remote-validation:
+    command: com.okta.spring.tests.common.servlet.code.BasicRedirectCodeFlowApplication
+    args:
+    - --server.port=${applicationPort}
+    - --okta.oauth2.issuer=https://localhost:${mockHttpsPort}/oauth2/default
+    - --okta.oauth2.clientId=OOICU812
+    - --okta.oauth2.scopes=offline_access
+    - --okta.oauth2.redirectUri=/authorization-code/callback
+    - --server.servlet.session.tracking-modes=cookie
+    protectedPath: /oauth2/authorization/okta
