+++ integration-tests/oauth2-servlet/src/test/resources/versions.properties
