+++ src/findbugs/findbugs-exclude.xml
@@ -53,7 +53,7 @@
 
     <Match>
         <!-- CSRF is disabled in the ITs to make testing RP-initiated logout easier -->
-        <Package name="com.okta.spring.tests.common.basic.code"/>
+        <Package name="com.okta.spring.tests.common.servlet.code"/>
         <Bug pattern="SPRING_CSRF_PROTECTION_DISABLED"/>
     </Match>
